Addon to track Misprinted Coins, Quest completions for free heiloom upgrades, and 1st heroic end boss kill of the day.
This addon allows the creation of a per toon database for tracking
Heirloom upgrade token quests that are tied to Misprinted Draenic Coins.

Also provides a DB display for quickly viewing progress in a frame on
multiple toons and servers at once.

Now has ability to track event currencies and items that allow
purchasing of Heirloom Upgrade tokens.

Eventually want to get to where can display if a toon has completed
their 1st random heroic a day as well.